extract trait
